+++ CHANGELOG.md
@@ -16,9 +16,11 @@ and other library components
 + now Query contains only SQL and parameters
 + ConnectionInterface no longer be responsible for fetching data
 + prefix for service's names
++ now one instance UnderQuery is associated with only one connection
 
 ### Removed
 + TableComponentTrait (logic moved to AbstractQueryBuilder)
++ ConnectionManagerInterface
 
 0.5.0 [2016-10-26]
 -----------------